pyproject.toml
[project]
name = "estudos_python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
   "djangorestframework>=3.16.0",
   "pytest>=8.4.1",
   "python-dotenv>=1.1.1",
]
